Add VantagePro2.RAIN_COLLECTOR_UNITS will have one of the following values: 0x00 = 0.01 in, 0x10 = 0.2 mm, or 0x20 = 0.1 mm. Default rain_collector_unit = 0.01 inch.
Vantage Pro2 incorporates a tipping-bucket rain collector in the ISS that
measures 0.01'' for each tip of the bucket. A metric adapter can be installed to
measure 0.2 mm for each tip of the bucket. Your station logs rain data in the same
units it is measured in and converts the logged totals into the selected display
units (inches or millimeters) at the time it is displayed. Converting at display
time reduces possible compounded rounding errors over time.